chore: move snapshot status management completely to snapshot worker
