+++ snapshotter/utils/snapshot_worker.py
@@ -1,4 +1,7 @@
 import importlib
+import json
+import time
+import asyncio
 from typing import Optional
 
 from ipfs_client.main import AsyncIPFSClient
@@ -6,9 +9,15 @@ from ipfs_client.main import AsyncIPFSClientSingleton
 
 from snapshotter.settings.config import projects_config
 from snapshotter.settings.config import settings
+from snapshotter.utils.callback_helpers import send_failure_notifications_async
+from snapshotter.utils.callback_helpers import send_telegram_notification_async
 from snapshotter.utils.data_utils import get_snapshot_submision_window
 from snapshotter.utils.generic_worker import GenericAsyncWorker
+from snapshotter.utils.models.data_models import SnapshotterIssue
+from snapshotter.utils.models.data_models import SnapshotterReportState
+from snapshotter.utils.models.data_models import SnapshotterStatus
 from snapshotter.utils.models.message_models import SnapshotProcessMessage
+from snapshotter.utils.models.message_models import TelegramSnapshotterReportMessage
 
 
 class SnapshotAsyncWorker(GenericAsyncWorker):
@@ -31,6 +40,7 @@ class SnapshotAsyncWorker(GenericAsyncWorker):
             task_type = project_config.project_type
             self._task_types.append(task_type)
         self._submission_window = 0
+        self.status = SnapshotterStatus(projects=[])
 
     def _gen_project_id(self, task_type: str, data_source: Optional[str] = None, primary_data_source: Optional[str] = None):
         """
@@ -90,17 +100,8 @@ class SnapshotAsyncWorker(GenericAsyncWorker):
                 'Exception processing callback for epoch: {}, Error: {},'
                 'sending failure notifications', msg_obj, e,
             )
+            raise
 
-            self.status.totalMissedSubmissions += 1
-            self.status.consecutiveMissedSubmissions += 1
-
-            await self._send_failure_notifications(
-                error=e,
-                epoch_id=str(msg_obj.epochId),
-                project_id=self._gen_project_id(
-                    task_type=task_type,
-                ),
-            )
         else:
 
             if not snapshots:
@@ -116,18 +117,24 @@ class SnapshotAsyncWorker(GenericAsyncWorker):
                     primary_data_source = None
                 else:
                     primary_data_source, data_source = data_sources
-
-                project_id = self._gen_project_id(
-                    task_type=task_type, data_source=data_source, primary_data_source=primary_data_source,
-                )
-                await self._commit_payload(
-                    task_type=task_type,
-                    _ipfs_writer_client=self._ipfs_writer_client,
-                    project_id=project_id,
-                    epoch=msg_obj,
-                    snapshot=snapshot,
-                    storage_flag=settings.web3storage.upload_snapshots,
-                )
+                    project_id = self._gen_project_id(
+                        task_type=task_type, data_source=data_source, primary_data_source=primary_data_source,
+                    )
+                try:
+                    await self._commit_payload(
+                        task_type=task_type,
+                        _ipfs_writer_client=self._ipfs_writer_client,
+                        project_id=project_id,
+                        epoch=msg_obj,
+                        snapshot=snapshot,
+                        storage_flag=settings.web3storage.upload_snapshots,
+                    )
+                except Exception as e:
+                    self.logger.opt(exception=True).error(
+                        'Exception committing snapshot payload for epoch: {}, Error: {},'
+                        'sending failure notifications', msg_obj, e,
+                    )
+                    raise
 
     async def process_task(self, msg_obj: SnapshotProcessMessage, task_type: str, preloader_results: dict):
         """
@@ -164,11 +171,28 @@ class SnapshotAsyncWorker(GenericAsyncWorker):
             task_type, msg_obj,
         )
 
-        await self._process(
-            msg_obj=msg_obj,
-            task_type=task_type,
-            preloader_results=preloader_results,
-        )
+        try:
+            await self._process(
+                msg_obj=msg_obj,
+                task_type=task_type,
+                preloader_results=preloader_results,
+            )
+        except Exception as e:
+            self.logger.error(f"Missed snapshot for epoch: {msg_obj.epochId}, task type: {task_type} - Error: {e}")
+            self.status.totalMissedSubmissions += 1
+            self.status.consecutiveMissedSubmissions += 1
+
+            await self._send_failure_notifications(
+                error=e,
+                epoch_id=str(msg_obj.epochId),
+                project_id=self._gen_project_id(
+                    task_type=task_type,
+                ),
+            )
+        else:
+            # reset consecutive missed snapshots counter
+            self.status.consecutiveMissedSubmissions = 0
+            self.status.totalSuccessfulSubmissions += 1
 
     async def _init_project_calculation_mapping(self):
         """
@@ -212,3 +236,50 @@ class SnapshotAsyncWorker(GenericAsyncWorker):
             await self._init_project_calculation_mapping()
             await self._init_ipfs_client()
             await self.init()
+
+    async def _send_failure_notifications(
+        self,
+        error: Exception,
+        epoch_id: str,
+        project_id: str,
+    ):
+        """
+        Sends failure notifications for missed snapshots.
+
+        Args:
+            error (Exception): The error that occurred.
+            epoch_id (str): The ID of the epoch that missed the snapshot.
+            project_id (str): The ID of the project that missed the snapshot.
+        """
+        notification_message = SnapshotterIssue(
+            instanceID=settings.instance_id,
+            issueType=SnapshotterReportState.MISSED_SNAPSHOT.value,
+            projectID=project_id,
+            epochId=str(epoch_id),
+            timeOfReporting=str(time.time()),
+            extra=json.dumps({'issueDetails': f'Error : {error}'}),
+        )
+
+        telegram_message = TelegramSnapshotterReportMessage(
+            chatId=settings.reporting.telegram_chat_id,
+            slotId=settings.slot_id,
+            issue=notification_message,
+            status=self.status,
+        )
+
+        tasks = [
+            asyncio.create_task(
+                send_failure_notifications_async(
+                    client=self._reporting_httpx_client,
+                    message=notification_message,
+                ),
+            ),
+            asyncio.create_task(
+                send_telegram_notification_async(
+                    client=self._telegram_httpx_client,
+                    message=telegram_message,
+                ),
+            ),
+        ]
+
+        await asyncio.gather(*tasks)
